Remove redundant docs parsing files
