+++ src/document/document_processor.ts
@@ -1,0 +1,1 @@
+export class DocumentProcessor {}
